requirements.txt
python-dotenv~=1.0.1
langchain~=0.1.14
pyspark~=3.4.2
pypdf2~=3.0.1